+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.5.2] - 2024-XX-XX
 - Fix issue [#336](https://github.com/intersystems/language-server/issues/336): Add intellisense for variables set to the returned value of a method, or a property, with a declared type
 - Fix issue [#337](https://github.com/intersystems/language-server/issues/337): Improve Hover headers
+- Fix issue [#338](https://github.com/intersystems/language-server/issues/338): Add intellisense for macros in Embedded SQL
 
 ## [2.5.1] - 2024-07-09
 - Fix issue [#328](https://github.com/intersystems/language-server/issues/328): Fix namespace detection for Diagnostic computation